Extract warnings for backup tasks
Previously the db task was unique in that it had special warnings.
Encode these warnings as part of the task interface so that warning
handling can be centralised. This means the db task can be handled in
the exact same way as every other task.

A spec had to be removed because all the methods that would output the
time have been stubbed. This had only worked before because db was
handled separately.