Backport `database_tasks:` from Rails 7.x
For some configurations of databases we should skip
creating, dropping or running migrations.

In particular this applies to Phase 4, where we should
not create, drop or migrate the `ci:` if it points to
the same database as `main:`.

This is well supported in Rails 7. This backports exact
same config to achieve exactly the same outcome.