Fix hanging specs when decomposed
This is due to a Rails bug causing where FactoryBot thread, and Puma
thread are stuck with each other waiting for the
load_interlock_aware_monitor lock.

The fix is to ensure no Puma threads ocurr when FactoryBot factories are
being created. Here it is done by moving the visits to within the spec,
after the factory creation.

Previously the visits were in the before hook, which meant async
requests ran concurrently with factories.