Add information about Cloudflare limits to instance_limits.md
The page https://docs.gitlab.com/ee/administration/instance_limits is
intended to be a comprehensive list of all application and rate limits
which apply to GitLab.  Information about Cloudflare CDN rate limits and
DDoS protection was previously not included.  This MR adds documentation
about those limits to this page.