Handle Jira Connect installation updates
The previous logic ignored installed hooks when the installation already
exists. This was because we assumed that the hook would only indicate
an App version update, but don't contain any data we need to process.

It turns out installed hooks for existing installation are aiming to
update those installations with a new auth token or base_url

This broke some active installations when auth tokens were rotated.

Changelog: fixed