Always pass data payloads as a hash for hooks
Web hook payloads must be plain hashes of data - they are serialized
when sending to sidekiq and when storing in the database.

To support this, we ensure that the payload is in
fact a hash in the execution service.

This adds a regression test
for https://gitlab.com/gitlab-org/gitlab/-/issues/353567, and simplifies
log_execution, since serveral of the arguments to this method are
fixed.

Changelog: fixed