Remove use_model_load_balancing feature flag
Previously these feature flags and environment variables were introduced
as a way to control the rollout of a new CI database. The new CI
database needs to be configured in `config/database.yml` but since
changing this required a full redeployment of all hosts it was not going
to be as easy to switch on/off as a feature flag so this feature flag
was setup such that we just ignored the values in `config/database.yml`
unless the feature flag was set.

Now that we've deployed "Phase 3" in production
https://gitlab.com/gitlab-com/gl-infra/production/-/issues/6513 we no
longer need these feature flags or environment variables to enable this
behaviour. Now we just use the values as set in the
`config/database.yml`.

Changelog: other