Drop CI builds triggered by blocked user
When a user is blocked, typically `Ci::DropPipelineService` will
cancel the builds. However, there's a chance that this service doesn't
pick up all the cancelable builds at the time it is run. To ensure
that builds aren't actually run for blocked user, we add it to a
precondition before we assign a runner.

Changelog: fixed

Relates to https://gitlab.com/gitlab-org/gitlab/-/issues/351294