Group access tokens cannot import projects
Import process always attempts at adding the importing user as a
direct member to the target project or group. This does not work
when the importing user is a group (or project) access token bot
account because of restrictions that prevent such a user from being
a direct member of any group other than the top level one.

This change adds a troubleshooting entry with the error
that is observed when a group access token is used to attempt
a project import.