Build deploy hook payload in-request
This change builds the deploy web hook payload in the request, rather
than delayed through `Deployments::HooksWorker`.

Once the payload is built, it is sent to the queue via `WebHookWorker`
and the webhook executes on the queue, as normal.

This change is behind a feature flag to allow us to test whether the
change will fix a problem for a particular customer who has some webhook
payloads containing a "status" of "running" when the deployment has
finished, when the status should be "succeeded".

https://gitlab.com/gitlab-org/gitlab/-/issues/355903