Remove Balsamiq File Preview
The Balsamiq File Preview is not utilized a lot. In fact it has been
broken since over a year on gitlab.com and on self managed instances[0].
The Balsamiq preview itself doesn't have a lot of value as it just
renders the preview of the first Balsamiq page and not all the pages.
We have received no user complaints whatsoever, meaning that likely
_noone_ relied on this feature.

This MR removes the previewer, as it is also riddled with tech debt, for
example shipping a asm.js compiled version of sql.js

[0]: https://gitlab.com/gitlab-org/gitlab/-/issues/285525

Changelog: removed