Refactor environment empty state to use scope
The empty state reported that you didn't have any environments
regardless of scope, which was a bit misleading if you had many
available environments and no stopped environments.

I have refactored the component to allow the scope to be injected so the
text can change depending on what scope the user is looking at.

I've also refactored the rest of the translations to follow more
up-to-date patterns, and rewritten the tests to utilize the testing
library's queries for finding things by role.

Changelog: fixed