Link to previous path when viewing blame on renamed files
On blame page, the "View blame prior to this change" button
fail if the file has been renamed.

To resolve that, when reading the raw blame from git,
the lines beginning with "previous" are read
to retrieve the previous path.

Also, the button has been removed for the initial commit
on the file.

Changelog: changed