Cleanup draft column data
In a prior, now reverted migration, we attempted to populate the draft
column based on the "draft" or "WIP" nature of the title, mirroring what
is used in `MergeRequest#draft?`. However, an error in the SQL regexp
caused us to grab an additional number of MRs and label them as "draft".

Changelog:other