Prevent loading wiki content for destroy and diff actions
In https://gitlab.com/gitlab-org/gitlab/-/merge_requests/81513 we
add a param to avoid loading the wiki page content. In
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/71753 we
avoid loading the content for the show and history actions.

But there are other actions where we don't need the wiki page
content and we still load it. These actions are `diff`
and `destroy`.

Changelog: performance