test: Allow setting up Praefect with per_repo strategy
At the moment Praefect in test is used with the in-memory election
strategy. We at Gitaly like to deprecate this mode and have tests
running with the default "per_repository" election strategy.

These changes make it possible to generate a Praefect configuration that
has all the setting to use this "per_repository" strategy.