Expose HTTP code during GitHub importer errors
This change adjusts the message returned to the user
when a GitHub import operation fails to include the
response HTTP code.

This is useful when the HTTP body is empty, such as
during the use of reverse proxies in accessing a
GitHub Enterprise instance over the internet.

The presence of a HTTP code can help establish if
a connection was at least made successfully and
the HTTP code can further be traced in their own
proxy or GitHub Enterprise logging for further
troubleshooting.

The logs already carry this information, but on
SaaS this is currently not visible due to
https://gitlab.com/gitlab-com/gl-infra/reliability/-/issues/15470

Changelog: other