Uses meta? to hide line comments and text
We were using discussable to hide the 0 text of lines on the
transformed diff that didn't have a mapping on the original
diff, but this wouldn't help with making them also not
notable.

Assigning this lines the type of `no-raw-mapping` and
adding this to the meta? solves both cases: lines that
don't have a mapping will both have no text and not be
notable