Detect projects built for Apple iOS platform
Like repository language detection feature, we'll try to detect if a
project is built for Apple iOS platform. This information will be used
in experiments to customize the experience (e.g. suggest setting up a
MacOS runners) of developers involved in such projects.

This detection will happen every time changes are pushed to the
project’s main branch and after a project is imported (using a project
template when creating a project is also an import). Detection is
executed in the background as a Sidekiq job at most once every hour
(using ExclusiveLease) and only if the project uses Swift or Objective-C
programming language.

If the project is detected as built for iOS we mark it by setting the
project’s corresponding ProjectSetting `target_platforms` attribute to
[‘ios’].

Changelog: added