When building from source, require nodeJS >= 14.15.0
The longterm support of node@12 will run out next month. Over a year ago
we started recommending node@14, as of recent we started using node@16
on our development machines and CI.

For a future eslint upgrade we need to drop support for node@12. Updates
to our minimum required node version can happen outside of the normal
breaking release cycle, as it is just needed for building from source
and is no runtime requirement.

- node@12 -> node@14 https://gitlab.com/gitlab-org/gitlab/-/merge_requests/52099
- node@14 -> node@16 https://gitlab.com/gitlab-org/gitlab/-/merge_requests/81372

Changelog: other