Include descendants when validating delegator overrides
It's possible that a method is only defined on a child class and not
on the parent class.

For example, `GroupLabel` and `ProjectLabel` define a `#subject` method
but `Label` doesn't, so we don't detect this method as being overriden
in `LabelPresenter`.

Instead of specifying all possible classes on the presenter, we can
just include the descendants in the validator.