Increase the visibility of the keyboard shortcut for Global Search
closes issue: https://gitlab.com/gitlab-org/gitlab/-/issues/342427/

This commit adds a keyboard shortcut visual cue to search the input field.
It is automatically hidden when input fields become active.
Visual cue adjusts to all possible color themes. It also overpowers
the clear button when there is input in the field but the field is not active.

Changelog: fixed