Fix showing closed issues as open in group/project issues list
An issue's `closedAt` property was checked to see whether it is
closed (to mimic the logic in the Haml template). However,
we added the `closed_at` database column 5 years ago and did
not backfill the data so issues older than this incorrectly
showed as open on the group/project issues list. This commit
fixes this by checking the issue's `state` property.

Behind feature flag `vue_issues_list` defaulted to off.