Persist issuable escalation statuses on issue update
Behind the :incident_escalations feature flag.

If a user updates the incident status or escalation policy
for an incident which does not yet have an IssuableEscalationStatus
record, we should create the record with the provided value.

If issue is associated with an alert, we want to pull in attributes
from the alert to populate the escalation policy of the incident.