Group analytics lookup group by full_path only
Before, it was looking up the group by id or full_path. That was buggy
in case the full path is only composed of digits, as in that case it was
assumed to be an id. Since the Groupi Activity Analytics API parameter
is called full_path, and documented to be the full path of the group,
there is no reason to support lookup by id.

Changelog: fixed
EE: true