Add cloud native index page
Cloud Native GitLab supports more than one ways to install GitLab.
Previously, we only had the charts, but we the addition of operator,
we need a central page to list all the cloud native ways to install
GitLab that will also serve as the landing page in the docs navbar.