Replace runners 'active' filters with 'paused'
This change removes the active/paused values from the status filter
to search for a runner and replaces them with a new "Paused" filter
with values Yes or No.

This new filter is an equivalent way to make this search, users that
visit URLs that use the deprecated status filters are redirected.

Changelog: changed