Update Product Planning Controller Urgencies
Controller actions loading HTML pages without applying filters
were given an urgency of medium (0.5s). However, they breach this
with almost identical frequency (circa 10%). This indicates there
is probably no information unique to the controller actions. It
also causes error budget spend which is not likely to be
attributable to the Rails stack.

This change reverts those urgencies to the default.