Update feature flag template to make staging-canary explicit
Part of https://gitlab.com/gitlab-com/gl-infra/delivery/-/issues/2280

With the change of the deployment pipeline for GitLab.com at

https://gitlab.com/groups/gitlab-com/gl-infra/-/epics/608

`staging-canary` is now the first environment deploy to, and this
the best place to test new features and feature flags.

This change updates the feature flag issue template to call this out
explicitly so developers now that is where they should be testing
their feature flags before production.