Remove pending builds from the queue on conflict
When the new denormalized pending builds table is enabled,
database-level consistency guarantees are reduced because of the
denormalization.

Here we are improving our resiliency mechanisms by properly reconciling
the database state in case of a conflicting data being detected. We
remove a queuing entry of a build that has already been processed.

Changelog: fixed