Add checkbox to include tag note in release notes
When a tag is selected, its details are fetched. If it has an included
message (it is an Annotated Tag), and the new checkbox is selected, the
tag notes are included in the following format:

```
release notes

### Tag Notes

annotated tag message
```

When editing a release, as the tag notes could already be included, the
option to include it is removed.