Add ability to pre/post-date audit events
When a sidekiq job creates an audit event
during its running, it'll create the event
when the job was executed, not enqueued.

When this happens, we should allow the creating
service to override the value.

This commit adds the ability to set a specific
time for when an audit event was created. By default
this will default to Time.zone.now (the current
behaviour) so there is no breaking change.

Changelog: fixed