Start storing change date for max_seats_used
We want to be able to efficiently determine what date a subscription had
its max_seats_used value updated so that we can determine whether or not
we need to inform the user of our overage policy.

Add this as a column to the Subscription model so that we do not have to
calculate this from the group's billable members each page load.

Issue: https://gitlab.com/gitlab-org/gitlab/-/issues/348481
Changelog: added
MR: https://gitlab.com/gitlab-org/gitlab/-/merge_requests/84913