Collect included files on CI pipeline creation
We want to show included files in the CI Linter page. To do that,
backend needs to send this information to frontend via the ciConfig
result.

In this commit, we only collect this data in a metadata method of
CI Config. In the next iteration, we'll pass this information to
the GraphQL endpoint.