Do not cache attention requests count
In https://gitlab.com/gitlab-org/gitlab/-/issues/357385, it was
reported that the cached count is inaccurate with the actual results.

In https://gitlab.com/gitlab-org/gitlab/-/merge_requests/84140, the
count query has been improved (from ~400ms down to ~14ms).

Since the query is performing better, we don't need to cache the
count.

This is behind `uncached_mr_attention_requests_count` feature flag.
The invalidation calls are not behind the feature flag to keep it
simple. When we decide to remove the feature flag, we can also
remove the calls for `User#invalidate_attention_requested_count`.