Fix repository analytics ref dropdown
The ref dropdown on the repo analytics was adding the
path suffix twice due to a JS fix made for another page
that uses the same dropdown. This change fixes the
issue for the repo analytics page while also keeping
the ability for other pages to use the previous fix.

Changelog: fixed