Switch from respond_with_navigational to redirect_to
This replaces respond_with_navigational with redirect_to when
redirecting to the sign-in form after a failed sign-in attempts.

There seems to be an issue with respond_with_navigational that prevents
feature flags from being pushed to the frontend. This issue does not
seem to happen when using redirect_to.

Changelog: fixed