banzai: Get Custom Emoji from group as well
Not all markdown fields are related to a project. Some, like those in
Epics, might be related to a group. So in addition to finding the
containing namespace through the project, also get try to get it from
the group or namespace.

Because we moved the responsibility of looking up the root namespace to
the scope #for_resource in CustomEmoji, we can rely on that.

This fixes custom emoji in Epic descriptions and notes.

Changelog: fixed