Wait until ArkoseLabs token is set before resubmitting
This fixes the ArkoseLabs suppressed challenge logic to ensure that we
wait until the token input has been hydrated before re-submitting the
form.
We previously relied on the onSuppress callback which is called too
early for the token to be set. We can instead rely on the onCompleted
callback and check the response to determine whether the challenge was
suppressed. We then wait a tick before submitting the form to make sure
the DOM is up-to-date when making the POST request.

Changelog: fixed