Make OmniAuth initializer return Geo proxied URL when it exists
OmniAuth full_host supports a proc, this makes it return a proc
that changes based on the env to verify whether it's a proxied
Geo request or not.

For regular requests, or non-EE requests, this keeps returning
the base URL, while proxied Geo requests return the URL of the
proxying site - to get SSO working with separate URLs.

EE: true
Changelog: changed