Add namespace settings attributes pipeline to group bulk imports
Add a new pipeline that migrates specific namespace settings attributes
from the exported NDJSON namespace_settings.ndjson. This commit also
adds this NDJSON file to the list of files to be generated during the
group export.

This NDJSON is not imported when importing the group from a file. It is
only used in the bulk import.

Changelog: added
EE: true