Expose url on discussion award emoji
The frontend fetches discussions through a JSON endpoint:
discussions.json. This endpoint also provides the awarded emoji for each
comment, but those did not include the url for those emoji images (when
available). The frontend code is already capable to deal with custom
awarded emoji since [1], and it is working for custom award emoji on the
issue itself. But for comments it was missing the url of the image.

With this change the url is also provided for award emoji presented
through the discussions.json REST endpoint.

1. 8f8eb1a844f (awardEmoji: Display custom emoji, 2022-03-14)

Changelog: fixed