Fix default job_arguments for batched migrations
The `batched_background_migrations` table previously had an incorrect
default value for the jsonb column `job_arguments`.

The incorrect default was '"[]"', but should have been '[]' (to be an
actual empty JSON array, rather than a string of an empty array).

This change fixes the default value, and updates any existing records
with empty argument lists to have the correct value as well.

Changelog: changed