Use deferrable constraint to bulk cadence update
Iterations cannot have overlapping dates within an iteration cadence.
We validate and enforce this at the appliation and DB level
respectively with AR validations and a PG exclusion constraint.

When updating iteration dates in bulk, we want to sometimes delay
the enforcement of the PG constraint until the end of a transaction.
This commit adds "DEFERRABLE" option to the exiting exclusion
constraint.

Changelog: changed