Remove support for uses_legacy_database_config
This was planned to to be removed in `15.0`. This setting in
`config/database.yml` (or inferred from the structure of
`config/database.yml`) allowed GitLab installations (from source) to
stick to the old "flat" structure. Since we've already updated Omnibus
installations and GDK to no longer generate this flat structure there
would only be installations from source that are affected by this change
and they should have already been receiving deprecation warnings for
some months. This change will ensure they see an error if they don't
have a top level `main` key in their `config/database.yml` and all they
need to do is nest the config under `main` to fix the problem.

Changelog: removed